+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.3.2] - 2023-08-22
+
+### Changed
+ * Use new version of toshi-hazard-store with faster, cheaper queries
+ * Update demo configs and SRM logic tree
+
+### Added
+ * Transpower locations
+
 ## [0.3.1] - 2023-08-11
 
 ### Changed